chore(base): Move `RoomInfoNotableUpdate*` inside `room_info`.
This patch moves the `RoomInfoNotableUpdate` and
`RoomInfoNotableUpdateReasons` types inside the `room_info` module.
